Return null as metadata object if all fields are nulls.
If we see nulls as metadata fields in the database, it's impossible to parse metadata back, because it's impossible to make a difference between two cases:
a) Use null as metadata
b) Use object with nulls in all fields, e.g. User(name=null, age=null, phone=null)

In both cases we will get [null, null, null] in the database, so, when we try to parse it back, we just return null as a safe choice. One more reason – there isn't a meaningful difference between 'null' and object 'User(name=null, age=null, phone=null)'